+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/Category.java
@@ -1,7 +1,7 @@
-package guru.qa.niffler.jupiter;
+package guru.qa.niffler.jupiter.annotation;
 
-import com.github.javafaker.Bool;
-import guru.qa.niffler.model.CurrencyValues;
+import guru.qa.niffler.jupiter.extension.CategoryResolverExtension;
+import guru.qa.niffler.jupiter.extension.CreateCategoryExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.lang.annotation.ElementType;
